EAMxx: fix usage of integer timestep in Homme interface

+++ components/eamxx/src/dynamics/homme/atmosphere_dynamics.hpp
@@ -45,8 +45,8 @@ public:
   // Cuda requires methods enclosing __device__ lambda's to be public
 protected:
 #endif
-  void homme_pre_process (const int dt);
-  void homme_post_process (const int dt);
+  void homme_pre_process (const double dt);
+  void homme_post_process (const double dt);
 
 #ifndef KOKKOS_ENABLE_CUDA
   // Cuda requires methods enclosing __device__ lambda's to be public
